Fixed some bugs ("games *","_ games")
Fixed errors inputting "games *" and "_ games", this was because I
didn't link the .csv files properly.

Also fixed the condition tags as I used them incorrectly.

Added templates for user to ask the bot what their own name or age is,
if this is known to the bot it will reply.