Added ability to send user link to games
Added ability for the bot to ask users favourite genre of game, then
their favourite game in that genre, if they own it and if not if they'd
like to buy it.
If yes it links them to the game on amazon.
Currently deals with 2 games of each genre.
Moved "what is my name" and "how old am i" out of their respective
topics. So they can be asked without the bot knowing their age or name
now and the reply is "I dont know this yet".
Also added ability for ask the bot what gender the user is.